+++ WinAPI/CTileMap.cpp
@@ -62,8 +62,8 @@ void CTileMap::GetEndIdxOfRectArea(int** _grid, int _startX, int _startY, int& _
 	vector<CObject*>& tiles = pCurScene->GetObjects(E_GroupType::TILE);
 
 	while (true) {
-		for (int y = startY; y < m_iRow; ++y) {
-			for (int x = startX; x < m_iCol; ++x) {
+		for (int y = startY; y < (int)m_iRow; ++y) {
+			for (int x = startX; x < (int)m_iCol; ++x) {
 				CTile* pTile = (CTile*)tiles[m_iCol * y + x];
 				if (E_TileType::Wall == pTile->GetTileType()) {
 					
@@ -77,11 +77,11 @@ void CTileMap::OptimizationTileCollider()
 {
 	// 2차원 배열 동적할당.
 	int **grid = new int* [m_iRow];
-	for (int i = 0; i < m_iRow; ++i)
-		grid[i] = new int[m_iCol];
+	for (int i = 0; i < (int)m_iRow; ++i)
+		grid[i] = new int[(int)m_iCol];
 
 
-	for (int i = 0; i < m_iRow; ++i)
+	for (int i = 0; i < (int)m_iRow; ++i)
 		memset(grid[i], 0, sizeof(int) * m_iCol);
 
 	// grid type -> 0 : non visited, 1: visited
@@ -91,8 +91,8 @@ void CTileMap::OptimizationTileCollider()
 	int startPosY = 0;
 	int endPosX = 0;
 	int endPosY = 0;
-	for (int y = 0; y < m_iRow; ++y) {
-		for (int x = 0; x < m_iCol; ++x) {
+	for (int y = 0; y < (int)m_iRow; ++y) {
+		for (int x = 0; x < (int)m_iCol; ++x) {
 			endPosX = startPosX = x;
 			endPosY = startPosY = y;
 			if ((int)E_TileType::Wall == grid[startPosY][startPosX]) {
@@ -106,7 +106,7 @@ void CTileMap::OptimizationTileCollider()
 		}
 	}
 
-	for (int i = 0; i < m_iRow; ++i) {
+	for (int i = 0; i < (int)m_iRow; ++i) {
 		delete[] grid[i];
 	}
 	delete[] grid;
